Convert md5 hash to hexstring
The md5 hash for images introduced in https://github.com/dainst/arachnebackend/commit/902d81fbb5e14a88e8756c71418bbf67312407a2 has been altered to return a hex string. This is because Felix Kußmaul who created the ticket resulting in the initial introduction of the hash (see previous commit) asked for it to be in that format.