Improve Places-time data, with nullable values. Also better sorting by date.
Since dates can have empty or negative values, the standard data types are not
useful for this. Here Integer-class types are used, so that null-values can be
supported.

The places for an object are sorted by start-date, starting with the year. Data
without valid start-date are moved to the bottom of the list, so that valid data
comes first.

Also introduced is a new compound-date-data type in the XMLs. This is used for
dates split up into parts (as mentioned before). The usual pre- and suffix
stuff still works.