Cleanup logging config
Most importantly re-insert the ROLLING appender for default output.
This way it is no longer written to catalina.out, but the
arachedataservice.log instead.

Also introduced a special logfile for Image-Caching output, because
it was cluttering up the logfile making it hard to see any real issues.

Furthermode removed unnecessary backup file and comments.