Fix con10t-media-tree demo
Not 100% sure this is how it is supposed to work: Instead of providing entity IDs to objects, we have to provide entity IDs to images in order to display something.

I expected to provide object IDs and the widget to select the first image for each each object or something similar.