Add `raw` and `formatted` field nesting
For reproduktion we nest one group of raw data fields and one for the preformatted data fields. The latter are displayed in the frontend as usual, the former can be used to extract data more granularly with automated processes.

This is pretty ugly, but short of rewriting all the XML entity transformation scripts and the frontend, this is the quickest solution.